Initial creation of media functionality
Created the main media function that is the functionality behind
styled-components-media. Currently, the media function has the ability
to accept a string, a number and an array as an argument. If the media
function recieves a string or a number, it will output a media query with
the feature 'min-width'. For example, if you input an argument of `100`
or '100', the function will return '@media (min-width: 100px)'. The media
if a number is passed the default unit will be `px`. If an array is
passed, the first item of the array will be the feature, and the second
will be the value. For example, if `['max-height', 500]` is passed, the
function will return `@media (max-height: 500px)`. Also created a
`createMedia` function. By default, if no argument is passed, the media
function will return the feature 'min-width' and the value `768px`. The
createMedia enables the default to be overwritten and will return a new
custom media function with the new default value.